refactor(core): lift _run_trial behind a typed Conductor Protocol
`Orchestrator._run_trial` was a 621-line instance method that did
everything between "you have a trial to run" and "here's the TrialResult":
environment setup, runner registration, agent-loop invocation, grading,
artifact writing. The orchestrator both scheduled trials AND executed
them.

After this PR, the orchestrator is a scheduler (run queue, retry policy,
result aggregation, lease loop). A Conductor is the executor (the
per-trial work). The orchestrator calls conductor.run(...) and gets a
TrialResult back; everything inside that call is the Conductor's concern.

Phase 1's seam-definition arc is complete after this PR: every plane
(control / trial / data / runtime / execution) now has a typed Protocol
with at least two implementations. Phase 2 (out-of-process trials,
remote conductor, distributed execution) becomes possible — shipping a
RemoteConductor is now "implement the Protocol against a gRPC client"
rather than "rewrite half the orchestrator."

Cut-and-paste discipline: the body of _run_trial moved verbatim into
InProcessConductor.run(); the three private helpers (_build_system_prompt,
_build_judge_messages_json, _serialize_model_config) and the module-
level _build_resolved_block (only consumer was _serialize_model_config)
moved with it. The orchestrator file shrinks by ~870 lines net.

The Conductor Protocol takes the 12-arg signature _run_trial already
had — extraction preserves it verbatim. A follow-up could collapse to
a TrialSpec-shaped input, but mixing extraction and reshaping in one
PR is unnecessary risk.

The Orchestrator.__init__ accepts an optional conductor_factory kwarg.
Default factory builds an InProcessConductor wired against the
orchestrator's per-run dependencies (adapter, artifact_writer, config,
logger, verbose, strict). Tests can inject an InMemoryConductor whose
call_log records every run() invocation and whose trajectory_factory
drives any success/failure/timeout scenario.

ADR-0008 documents the decision, considered alternatives, the
naming-overlap scope note (TrialRunner in core/runner.py is the agent
loop *inside* a trial, not the Conductor), and follow-ups: refactor
_run_trial's internals, RemoteConductor concrete implementation,
collapse DockerRunnerAdapter, promote RunnerClient to its own Protocol,
collapse the 12-arg run() signature into a TrialSpec input.

Linked tickets: TECHDEL-424 (Jira) / #100 (GH issue).
